[TF] Bugfix to conv recreation logic in tf.keras.layers.Conv*
Prior to this, if Conv()/ConvXD() was called with two different batch sizes
it would recreate the conv op unnecessarily.

Now that we have the Conv op being called with possibly two different batch
*ranks*, the code actually breaks and raises a ValueError.

Instead, we only check the non-batch part of the input shapes (the self.rank+1
inner dims) to decide whether to recreate the conv op.

PiperOrigin-RevId: 315080669
Change-Id: Ie92c2dbac3cccab6e7eeb92e2e5d4cb24e63a38a