Move AddBuiltin and AddCustom out of the interface.
To get the desired code size reduction from the newer API for adding
builtin operators, we need the AddBuiltin function to be removed by the linker.

However, linkers currently have limited support for removing unused functions
that are virtual.

This change moves the AddBuiltin function out of the interface (and AddCustom as
well for consistency).

https://stackoverflow.com/q/17433791
https://reviews.llvm.org/D63932
http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.kui0101a/armlink_bhcgeaid.htm

PiperOrigin-RevId: 315372426
Change-Id: I0dd1fc8a62dbe60076c7c847b43294ac273d0bad