Unify convert to constants logic.
Before this change there were two different code paths for dealing
with graph freezing for v1 and v2 graphs. They largely did the same
thing, but each path had and lacked capabilities the other had, and
each had its own bugs.

This change re-writes the previous v2 logic so it can cope with
session-based graphs and allow for the conversion of a subset of the
variables, and changes the previous convert_variables_to_constants
call to proxy into the new logic.

The new logic is built around a more "graphy" algorithm: variables are
converted to constants, and that conversion is then propagated through
the graph by following the graph edges. This hopefully makes it easier
to understand what is going on, and to change it later on.

More granular tests were added, in order to check that the right graph
manipulations were performed. In order to do that, some graph merging
infrastructure had to be created in the test.

PiperOrigin-RevId: 315497124
Change-Id: I3a33acc804b5dc9628c208df8fd1b7c59f906ddb