Do not rewrite std.return to lhlo.terminator when converting hlo to lhlo.
The lhlo.terminator is not legal in the places where the std.return operation is used. It is only used for regions of the lhlo dialect itself. As a next step, we need to replace existing uses of the lhlo.terminator operation with std.return as well, to align hlo and lhlo again.

PiperOrigin-RevId: 315674087
Change-Id: I64dcb3d9e07963db58c1cfeb680371873ad0a554