Parallel device: switch to manual threading to avoid synchronization issues with remote eager
Does not create new executors for each of its threads at the moment, so I'm not sure how this will interact with a user-level async setting. It'll be easy enough to create one sync executor per thread if that's a problem.

Requires some tweaks to TFE_OpAddAttrs to make it safe to call for different ops in different threads (due to protobuf arena destruction being global and not thread-safe). These changes avoid serialization+deserialization of attributes, so they make sense anyway.

PiperOrigin-RevId: 315990383
Change-Id: I211ceae1eb34e490b4d2a1df76226995c7ae7929