GCS File system: Remove tempfile in destructor regardless of status.
Otherwise this implementation leaves tempfiles around on failures, filling
up the temp filesystem: the Sync() call fails and Close() returns before the tempfile can be removed on destruction.

By removing the tempfile in the destructor, this mirrors the creation of the
tempfile in the constructor, and ensures it only happens once.

PiperOrigin-RevId: 316171772
Change-Id: I7a6ab9c417cd32b0e1fb46499b80f23bb476b6b6