Allow escaping return values in kernel lowering.
For simple cases where the launch only has function arguments and results as operands, the kernel signature will now be rewritten to have its operands ordered in the order of the function arguments followed by results.

This only works for simple functions without control flow or multiple launches and is intended for cases where the kernel is meant to be extracted. If the host-side is also used, signature rewriting should be disabled via the provided flag.

PiperOrigin-RevId: 316654295
Change-Id: Ie3e78b78bab2191610875acb79506e5986a821af