*** Reason for rollback ***
 CL316577397 prevents H2D prefetch overlapping for TF2 MLPerf Resnet model on V100x8
More details in b/159372996

  With the rollback, the prefetching overlapping is back, and the training speed also recovered.
*** Original change description ***

Add DT_BOOL support to GPU variable ops

This is a follow-on to PR #38848 & PR #39172 and resolves remaining ask
in Issue #35994. The original PR tried to add many variable ops on the
GPU including DT_BOOL. However, this caused testCondModifyBoolPred to
fail and thus the DT_BOOL type was removed. The reason for the test
failure is once DT_BOOL variables are supported on the GPU, we need to
ensure the switch ops are also updated to not have ho...

***

PiperOrigin-RevId: 317241338
Change-Id: Id7b7d79622e0537ccb677f081b487014ac4d2395