Fix TypeError when sparse.from_dense is called with tf.string
This PR tries to fix the issue raised in 40633 where
sparse.from_dense throws out TypeError with tf.string as input.

The issue was that from_dense uses `tf.constant(0, dtype)` to get
the zero value which fails on tf.string. This PR changes to `zeros_like`
to work with tf.string.

This PR fixes 40633.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>