Prepping for switch to ROCm 3.5+
Starting with ROCm 3.5 the underlying compiler used by hipcc will change from HCC to hip-clang. There will be a corresponding change in the HIP Runtime as well. This commit is part of a series which are intended to make the transition to ROCm 3.5+ easier.

This commit adds an alternative lookup path for `ld.lld` (since its location within the rocm install, will move in ROCm 3.5+).