[MLIR] This CL does several NFC changes: * Export the emitter in the header * Add EmitXXXOp to return the created op. * Change ::Run() to ::Initialize(), to make it possible for the LHLO FuncOp to be created but left empty.
The motivation is that XLA/GPU would be able to convert a single HLO op (through EmitSortOp in this CL) to a single LHLO op, without relying on the complete support of all ops. In this use pattern, there won't be a functioning LHLO graph, but supported ops can be used for XLA/GPU codegen.

PiperOrigin-RevId: 319137430
Change-Id: I5362ccb155af112bbf33c3bac9dc240f9227b1d4